Build config (#150)
* reqs separation

* fix reqs

* fix reqs

* without mlbox

* fix reqs

* add tf for 3.7

* fix reqs

* separate test directories

* naive separation

* check without b_reqs

* fix reqs

* remove examples

* refactor deps

* fix extra_reqs

* fix test

* extract benchmark

* remove mlbox

* rtd source packages on

* revert rtd

* fix extra_reqs
