--- benchmark/benchmark_model_types.py
@@ -1,10 +1,0 @@
-from enum import Enum
-
-
-class BenchmarkModelTypesEnum(Enum):
-    h2o = 'h2o',
-    tpot = 'tpot',
-    autokeras = 'AutoKeras',
-    mlbox = 'mlbox',
-    fedot = 'fedot',
-    baseline = 'baseline'
