--- benchmark/test_cases/penn_ml/datasets.csv
@@ -1,16 +1,0 @@
-dataset_names
-analcatdata_authorship
-201_pol
-analcatdata_bankruptcy
-207_autoPrice
-analcatdata_boxing1
-210_cloud
-hayes-roth
-616_fri_c4_500_50
-heart-c
-617_fri_c3_500_5
-heart-h
-618_fri_c3_1000_50
-heart-statlog
-620_fri_c1_1000_25
-cmc
