+++ examples/chain_tune.py
@@ -1,7 +1,7 @@
 import numpy as np
 from sklearn.metrics import roc_auc_score as roc_auc
 
-from benchmark.benchmark_utils import get_scoring_case_data_paths
+from cases.data.data_utils import get_scoring_case_data_paths
 from core.composer.chain import Chain
 from core.composer.node import PrimaryNode, SecondaryNode
 from core.models.data import InputData
