+++ extra_requirements.txt
@@ -1,0 +1,2 @@
+tpot==0.11.1
+h2o==3.28.1.2
