+++ requirements.txt
@@ -14,10 +14,7 @@ xgboost==1.0.1
 Pillow==7.1.0
 imageio==2.8.0
 pydot==1.4.1
-pmlb==1.0
-seaborn==0.9.0
-mlbox==0.8.4
-autokeras==1.0.2
-tensorflow==2.1.0
-tpot==0.11.1
-h2o==3.28.1.2
+seaborn==0.9.0; python_version < '3.7'
+tensorflow==2.1.0; python_version <= '3.7'
+tensorflow==2.3.0; python_version == '3.8'
+hyperopt==0.2.4
